Create ppp.sh
This script will initiate a PPP session and will present a fake login shell, but will not wait around for user input. The fake login shell is in place for compatibility with Trumpet Winsock 3.0, as it by default expects one, so we're fooling it to make it think that it's logging on. It will then proceed when it receives the expected printouts. The built-in dial-up connection in Windows 95/98/Me by default do not expect a login prompt unless specifically told to do so.

When Trumpet Winsock is in PPP mode, by default it will expect the following output after dialing the ISP's number and establishing a connection:

A username prompt, matched by Trumpet with the text “sername:”
A password prompt, matched by Trumpet with the text “ssword:”
A command prompt, matched by Trumpet with the character “>”
This script has been tested with the default installation of Trumpet Winsock 3.0 revision D with PPP mode switched on. This script has also been tested with the default dial-up utility of Windows 95 and Windows 98 with PPP enabled.

I've added a parameter to send an LCP echo to the client to test if the connection is still up. If the connection has abruptly been closed, pppd will know this by not receiving an echo reply, and will exit and relinquish control back to the vmodem.sh script. The only reason the timeout is in there, is because it seems like Trumpet Winsock 3.0 doesn't know how to tell pppd to terminate a PPP session from within a PPP session, and it will just hang up the call. As a result, pppd daemon will be left running indefinitely and won't ever give control back to vmodem. This is obviously not preferred, so LCP echo is added to let pppd know when the link has been cut. If you can think of better ways to accomplish this check, feel free to send tips on how to improve the script.